requirements.txt
langchain
langchain-community
chromadb
sentence-transformers
pdfplumber
python-dotenv
groq
streamlit